search-engine-pytorch/requirements.txt
pandas==1.3.4
sentence_transformers==2.2.0
torch==1.10.0+cu113
pillow==8.4.0
